`ItemTypeCollectionHStack` instead of iterating in the view itself. Use this for both Collections and Persons. Eventually, use this for Favorites.
