+++ Swiftfin tvOS/Views/ItemView/CollectionItemContentView.swift
@@ -26,17 +26,11 @@ extension ItemView {
                     .frame(height: UIScreen.main.bounds.height - 150)
                     .padding(.bottom, 50)
 
-                ForEach(viewModel.sections.elements, id: \.key) { element in
-                    if element.value.elements.isNotEmpty {
-                        PosterHStack(
-                            title: element.key.pluralDisplayTitle,
-                            type: .portrait,
-                            items: element.value.elements
-                        )
-                        .onSelect { item in
-                            router.route(to: .item(item: item))
-                        }
-                    }
+                ForEach(
+                    viewModel.sections.elements,
+                    id: \.key
+                ) { element in
+                    ItemTypeCollectionHStack(element: element)
                 }
 
                 if viewModel.similarItems.isNotEmpty {
